fix: Addressing feedback;
I've modified the filename to match the convention of `<port>.tera`. I've also
removed the unnecessary `flavor.colors.` prefix for colors. I've updated
the filenames of the generated files to use `identifier` instead of `name` with
a kebobcase & lowercase filename. Lastly, I've also updated the Comments color
to use `overlay0` instead of `subtext0`.